% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_complication_data}
\alias{pasteur_complication_data}
\title{pasteur complication_data}
\usage{
pasteur_complication_data()
}
\description{
complication_data function in here we would write a detailed description of this function - this is part of Jyoti's homework

UNZIPPING INSTRUCTIONS:
columns:"redcap_event_name", "visit_date", "complications", 
"first_complication_note", "another_complication_note", "diagnosis_date", 
"relatedness", "severity", "treatment_related", "new_or_previous_diagnosis", "form_notes"
column sep:"|"
row sep:";new_row: "
}
\examples{
\dontrun{
pasteur_complication_data()
}
}
